requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-jose[cryptography]
passlib
bcrypt
python-multipart
pydantic
pydantic[email]
python-dotenv
python-json-logger